(PC-20410) refacto(adaptOffersPlaylistParameters): move buildOfferGenreTypes and buildOfferGenreTypesValues closer to where it is used
