+++ src/libs/algolia/fetchAlgolia/fetchMultipleOffers/buildOfferGenreTypes.native.test.ts
@@ -1,5 +1,5 @@
 import { GenreType } from 'api/gen'
-import { buildOfferGenreTypes } from 'libs/search/utils/buildOfferGenreTypes'
+import { buildOfferGenreTypes } from 'libs/algolia/fetchAlgolia/fetchMultipleOffers/buildOfferGenreTypes'
 import { useGenreTypeMapping } from 'libs/subcategories/mappings'
 import { placeholderData } from 'libs/subcategories/placeholderData'
 import { renderHook } from 'tests/utils'
