+++ src/libs/algolia/fetchAlgolia/fetchMultipleOffers/buildOfferGenreTypes.ts
